feat/listProduct: move file component dropdown filter to folder components
