+++ src/pages/List Products/ListProductPromo.jsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import BookCard from "../../components/Book Card/BookCard";
 import Footer from "../Footer/Footer";
 import Header from "../Header/Header";
-import DropdownFilter from "./components/DropdownFilter";
+import DropdownFilter from "../../components/DropdownFilter/DropdownFilter";
 import axios from "axios";
 import { useParams } from "react-router";
